requirements.txt
argparse>=1.2.1
distribute>=0.6.24
gunicorn>=0.16.1
pymongo>=2.4.1
tornado>=2.4.1
wsgiref>=0.1.2
-e git+git://github.com/baden/mongodb-log.git#mongo-log
redis>=2.7.2
pyzmq>=2.2.0
mock=1.0
python-dateutil=2.1
hiredis>=0.1
coverage=3.6